+++ src/app/Balloon.App.Elasticsearch/Console/Elasticsearch.php
@@ -15,6 +15,8 @@ use Balloon\App\Elasticsearch\Job;
 use Balloon\Filesystem;
 use Balloon\Server;
 use GetOpt\GetOpt;
+use InvalidArgumentException;
+use MongoDB\Driver\Exception\ServerException;
 use Psr\Log\LoggerInterface;
 use TaskScheduler\Scheduler;
 
@@ -49,14 +51,22 @@ class Elasticsearch
     protected $fs;
 
     /**
+     * Bulk.
+     *
+     * @var int
+     */
+    protected $bulk = 200;
+
+    /**
      * Constructor.
      */
-    public function __construct(GetOpt $getopt, Server $server, Scheduler $scheduler, LoggerInterface $logger)
+    public function __construct(GetOpt $getopt, Server $server, Scheduler $scheduler, LoggerInterface $logger, array $config = [])
     {
         $this->logger = $logger;
         $this->getopt = $getopt;
         $this->fs = $server->getFilesystem();
         $this->scheduler = $scheduler;
+        $this->setOptions($config);
     }
 
     /**
@@ -70,18 +80,61 @@ class Elasticsearch
             return false;
         }
 
-        $this->logger->info('reindex elasticsearch requested, this is an async task, be sure you have enaugh workers running.', [
+        $total = $this->fs->countNodes();
+        $stack = [];
+        $done = 0;
+        $result = false;
+        $skip = $this->getopt->getOption('skip') | 0;
+        $this->bulk = $this->getopt->getOption('bulk') ? (int) $this->getopt->getOption('bulk') : $this->bulk;
+        $total -= $skip;
+
+        if ($total < 0) {
+            $total = 0;
+        }
+
+        $this->logger->info('reindex elasticsearch, nodes left: {total}/{total}', [
             'category' => get_class($this),
+            'total' => $total,
         ]);
 
-        foreach ($this->fs->findNodesByFilter([]) as $node) {
-            $this->scheduler->addJob(Job::class, [
-                'id' => $node->getId(),
-                'action' => Job::ACTION_CREATE,
-            ]);
+        $stack = [];
+        $done = 0;
+
+        while ($result === false) {
+            try {
+                $this->index($total, $skip);
+                $result = true;
+            } catch (ServerException $e) {
+                $skip -= $this->bulk;
+
+                $this->logger->error('cursor timeout captchered, restart from {skip}', [
+                    'category' => get_class($this),
+                    'exception' => $e,
+                    'skip' => $skip,
+                ]);
+            }
         }
 
         return true;
+    }
+
+    /**
+     * Set options.
+     */
+    public function setOptions(array $config = [])
+    {
+        foreach ($config as $key => $value) {
+            switch ($key) {
+                case 'bulk':
+                    $this->{$key} = (int) $value;
+
+                break;
+                default:
+                    throw new InvalidArgumentException('invalid option '.$key.' given');
+            }
+        }
+
+        return $this;
     }
 
     /**
@@ -99,6 +152,43 @@ class Elasticsearch
      */
     public static function getOptions(): array
     {
-        return [];
+        return [
+            \GetOpt\Option::create('b', 'bulk', \GetOpt\GetOpt::REQUIRED_ARGUMENT)->setDescription('Specify the bulk size [Default: 200]'),
+            \GetOpt\Option::create('s', 'skip', \GetOpt\GetOpt::REQUIRED_ARGUMENT)->setDescription('Instead start from node 0, you may skip nodes by specifing a different start index'),
+        ];
+    }
+
+    /**
+     * Index.
+     */
+    protected function index(int $total = 0, int &$done = 0): bool
+    {
+        $stack = [];
+
+        foreach ($this->fs->findNodesByFilter([], $done) as $node) {
+            $stack[] = $this->scheduler->addJob(Job::class, [
+                'id' => $node->getId(),
+                'action' => Job::ACTION_CREATE,
+            ]);
+
+            if (count($stack) >= $this->bulk) {
+                $this->logger->info('waiting for ['.$this->bulk.'] jobs to be finished', [
+                    'category' => get_class($this),
+                ]);
+
+                $done += $this->bulk;
+                $this->logger->info('reindex elasticsearch {percent}, nodes left: {done}/{total}', [
+                    'category' => get_class($this),
+                    'percent' => round($done / $total * 100, 1).'%',
+                    'total' => $total,
+                    'done' => $done,
+                ]);
+
+                $this->scheduler->waitFor($stack);
+                $stack = [];
+            }
+        }
+
+        return true;
     }
 }