+++ src/app/Balloon.App.Idp/Storage/Db.php
@@ -145,10 +145,16 @@ class Db extends OAuthMongoDB
      */
     public function setAccessToken($access_token, $client_id, $user_id, $expires, $scope = null)
     {
+        if ($expires === 0) {
+            $expires = null;
+        } else {
+            $expires = new UTCDateTime($expires * 1000);
+        }
+
         $token = [
             'access_token' => $access_token,
             'client_id' => $client_id,
-            'expires' => new UTCDateTime($expires * 1000),
+            'expires' => $expires,
             'user_id' => $user_id,
             'scope' => $scope,
             'adapter' => $this->adapter,
@@ -173,7 +179,11 @@ class Db extends OAuthMongoDB
      */
     public function setAuthorizationCode($code, $client_id, $user_id, $redirect_uri, $expires, $scope = null, $id_token = null)
     {
-        $expires = new UTCDateTime($expires * 1000);
+        if ($expires === 0) {
+            $expires = null;
+        } else {
+            $expires = new UTCDateTime($expires * 1000);
+        }
 
         return parent::setAuthorizationCode($code, $client_id, $user_id, $redirect_uri, $expires, $scope, $id_token);
     }
@@ -183,7 +193,11 @@ class Db extends OAuthMongoDB
      */
     public function setRefreshToken($refresh_token, $client_id, $user_id, $expires, $scope = null)
     {
-        $expires = new UTCDateTime($expires * 1000);
+        if ($expires === 0) {
+            $expires = null;
+        } else {
+            $expires = new UTCDateTime($expires * 1000);
+        }
 
         return parent::setRefreshToken($refresh_token, $client_id, $user_id, $expires, $scope);
     }
@@ -199,7 +213,9 @@ class Db extends OAuthMongoDB
             return false;
         }
 
-        $token['expires'] = $token['expires']->toDateTime()->format('U');
+        if ($token['expires'] !== null) {
+            $token['expires'] = $token['expires']->toDateTime()->format('U');
+        }
 
         return $token;
     }
@@ -217,7 +233,9 @@ class Db extends OAuthMongoDB
             return false;
         }
 
-        $code['expires'] = $code['expires']->toDateTime()->format('U');
+        if ($code['expires'] !== null) {
+            $code['expires'] = $code['expires']->toDateTime()->format('U');
+        }
 
         return $code;
     }
@@ -233,7 +251,9 @@ class Db extends OAuthMongoDB
             return false;
         }
 
-        $token['expires'] = $token['expires']->toDateTime()->format('U');
+        if ($token['expires'] !== null) {
+            $token['expires'] = $token['expires']->toDateTime()->format('U');
+        }
 
         return $token;
     }