+++ src/app/Balloon.App.Office/.container.config.php
@@ -10,23 +10,6 @@ return [
     ClientInterface::class => [
         'use' => Client::class,
     ],
-    AbstractBootstrap::class => [
-        'calls' => [
-            'Balloon.App.Office' => [
-                'method' => 'inject',
-                'arguments' => ['object' => '{'.Http::class.'}']
-            ],
-        ]
-    ],
-    Http::class => [
-        'arguments' => [
-            'config' => [
-                'loleaflet' => "{ENV(BALLOON_OFFICE_URI,https://localhost:9980/loleaflet)}/dist/loleaflet.html",
-                'wopi_url' => "{ENV(BALLOON_WOPI_URL,https://localhost)}",
-                'token_ttl' => 3600
-            ]
-        ]
-    ],
     Converter::class => [
         'calls' => [
             Office::class => [
@@ -35,18 +18,4 @@ return [
             ]
         ],
     ],
-    Office::class => [
-        'services' => [
-            ClientInterface::class => [
-                'arguments' => [
-                    'config' => [
-                        'base_uri' => '{ENV(BALLOON_LIBREOFFICE_URL,https://libreoffice:9980)}',
-                        'connect_timeout' => 3,
-                        'timeout' => 10,
-                        'verify' => false
-                    ]
-                ]
-            ]
-        ]
-    ]
 ];