+++ src/lib/Filesystem/Node/File.php
@@ -371,6 +371,7 @@ class File extends AbstractNode implements IFile
             'version' => $this->version,
             'parent' => $this->parent,
             'acl' => $this->acl,
+            'lock' => $this->lock,
             'app' => $this->app,
             'meta' => $this->meta,
             'mime' => $this->mime,